#278: Extract more filter functions into helper file
